refactor(ConversationCells): remove file deletion and restoration handlers, integrate them into CellsTableRowOptions for better modularity
