+++ src/script/components/Conversation/ConversationCells/CellsTable/CellsTableColumns/CellsTableColumns.tsx
@@ -36,13 +36,9 @@ const columnHelper = createColumnHelper<CellItem>();
 export const getCellsTableColumns = ({
   cellsRepository,
   conversationQualifiedId,
-  onDeleteNode,
-  onRestoreNode,
 }: {
   cellsRepository: CellsRepository;
   conversationQualifiedId: QualifiedId;
-  onDeleteNode: ({uuid, permanently}: {uuid: string; permanently?: boolean}) => void;
-  onRestoreNode: ({uuid}: {uuid: string}) => void;
 }) => [
   columnHelper.accessor('name', {
     header: t('cellsGlobalView.tableRowName'),
@@ -75,8 +71,6 @@ export const getCellsTableColumns = ({
       return (
         <CellsTableRowOptions
           node={info.row.original}
-          onDelete={onDeleteNode}
-          onRestoreNode={onRestoreNode}
           cellsRepository={cellsRepository}
           conversationQualifiedId={conversationQualifiedId}
         />
